requirements.txt
# Django & DRF
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Authentication
djangorestframework-simplejwt==5.3.0

# Image Processing
Pillow==10.0.0

# API Documentation
drf-yasg==1.21.7

# Database
psycopg2-binary==2.9.9

# Environment Variables
python-dotenv==1.0.0

# Password Hashing
bcrypt==4.0.1

# CORS
django-cors-headers==4.3.0

# Validation
django-phonenumber-field==7.1.0
phonenumbers==8.13.19

# Development Tools
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.0
pytest-django==4.5.2
factory-boy==3.3.0

# Code Formatting
black==23.7.0
flake8==6.1.0

# Production Server
gunicorn==21.2.0
whitenoise==6.5.0